combine Slideshow and Preview into one
