+++ src/components/Preview/useSyncSlideshow.ts
